Isolate template types
