+++ src/models/template/raw-template.ts
@@ -1,0 +1,29 @@
+/*************************************************************************
+ * Copyright 2020 Gravwell, Inc. All rights reserved.
+ * Contact: <legal@gravwell.io>
+ *
+ * This software may be modified and distributed under the terms of the
+ * MIT license. See the LICENSE file for details.
+ **************************************************************************/
+
+import { RawNumericID, RawUUID } from '../../value-objects';
+
+export interface RawTemplate {
+	GUID: RawUUID;
+	ThingUUID: RawUUID;
+	UID: RawNumericID;
+	GIDs: null | Array<RawNumericID>;
+	Global: boolean;
+	Name: string;
+	Description: string; // Empty string is null
+	Updated: string; // Timestamp
+	Contents: {
+		query: string;
+		variable: string;
+		variableLabel: string;
+		variableDescription: string | null;
+		required: boolean;
+		testValue: string | null;
+	};
+	Labels: null | Array<string>;
+}
